iiod: Move interpreter function to interpreter.c
Move the interpreter() function, along with the implementations for the
.readfd / .writefd callbacks, from ops.c to a new interpreter.c file.

This allows to decouple the interpreter() function from the old v0.x
protocol implemented in ops.c.
